Completion of tasks number 7, 8, 20. Links below. Config modal now accepts row and column from user. Back-end now builds board to spec, front-end displays board to spec.
Note: New means of creating board on back-end was implemented in this
commit. I ran into issues with the recursive call and the stack when
working with very large boards. Now a while loop is implemented and
each subsequent loop increases likely-hood of a cell being a mine. This
resolved my issues with very large boards.

https://blakerieschick.orangescrum.com/dashboard#/details/0bf96b74dfd0faed3cccf128fc22f3c4
https://blakerieschick.orangescrum.com/dashboard#/details/080885f69d6e62c6fe6d7f69b9553d38
https://blakerieschick.orangescrum.com/dashboard#/details/d33f3ca0e8e2ee5d0dc905884b05cbb6