Completion of task number 23, link below.
https://blakerieschick.orangescrum.com/dashboard#/details/c54050f97a3796ba14b06c7f72127f43

Previously, showing and hiding of each error label in the bad config
modal was done at the same time as determining if there was bad input.
This worked fine with only one board input from user, but with row and
column input the previous logic did not scale and was overwriting the
showing and hiding of the bad board input label. Now this is rewritten
to determine whether or not to show or hide each label until after
checking each value for good input.