Getting preset config logic from previous commit to work with high score logic.
Calling presetConfig now handles setting a preset_index variable on the
board, as well as board generation and calling for game to start with
modal manager. This preset_index is used when user wins and we check if
user is using preset, if so, the index tells us which one so we can see
if they have a high score and where to store it locally before pushing
store.